Migrate events to chai-kefir style
Pushes the events through the component factory.
